Work-around materialsystem.so asking for SDLMgrInterface001.
The materialsystem shared object currently asks our fileSystemFactory for
the SDLMgrInterface001 interface, which it doesn't provide. After that,
materialsystem self-destructs because it can't load our shaderapiempty
shared object. However, materialsystem doesn't really need that interface
so we'll simply set up a redirector that catches the request and returns
a dummy object it won't really use.