Source Engine Co-Operative Base Modification 2013.
This truly is the final commit.
Forgot to add the latest project files to last commit (client is
unchanged, however server has the old portal code files removed in
favour of the newly commited ones.
Having made sure these project files did compile, decided to commit the
newly built .dll files just so everyone is on the same page.
Also a new readme to reflect that this is the final original author
commit.
Good luck and happy modding!