Fix segfault in NDebugOverlay::Circle when debugoverlay is nullptr
The debugoverlay interface global is initialized to nullptr in the dedicated server. All of the NDebugOverlay functions which call into debugoverlay make sure to check whether it's nullptr before calling, with the exception of NDebugOverlay::Circle, which does no such check. This means that certain debug convars can cause the dedicated server to crash (e.g. nb_debug).