fix: scout balls computing critical hits incorrectly
scout balls were computing crits in their own entity think
outside of the player command, which made the random
generation invalid. this caused random crits to be computed
incorrectly, and also allowed for an exploit where you could
store a random crit with the ball and permanently crit every ball thereafter.

there is no way to influence the crit between the secondary
attack command and the ball think running, because the primary
attack time is delayed by 0.25 seconds, which is longer than the
launch delay. for safety, tf_scout_bat_launch_delay could be limited
to 0.25 at most but i'm not sure if this is required.