Fix medigun prediction errors
All of the code which determines how much ubercharge you get on a specific tick was server only, causing the medigun to constantly cause prediction errors unless you are in a situation where you are getting the default ubercharge per tick (stock medigun, damaged patient, no critheals).

This fixes that by making all possible server code to also be on the client.
The only exception to this is an unused weapon attribute which would modify ubercharge rate as long as you weren't in a respawn room. The functions it calls are server only, and would require changes to how respawn room entities are handled on the client to fix.