Initialize & deinitialize memory in CUtlMemory
Currently, CUtlMemory doesn't initialize anything it receives from malloc(), realloc(), and it doesn't de-initialize memory before calling free(). This makes it incompatible with the C++ STL, and requires anyone using these constructs to pay very close attention if they want to use RAII features. (CUtlMemory also memcpy's the underlying memory, when a swap would probably be more consistent)

This commit modifies the plain CUtlMemory to ensure that all newly-allocated memory is initialized using the in-place constructor and destructed using the destructor.

Why is this significant? During copy or move operations, the STL assumes that original object is always assumed to be in an assignable and destructible state. If you assign a C++ smart pointer to uninitialized memory, it will try to destruct the uninitialized memory first, treating it as if it was a valid smart pointer. This commit ensures that the memory will always be, at the very least, assignable and destructible.