Update tf_bot_attack_flag_defenders.cpp
Inside of function Update:
The absence of a flag on the map will now cause the Update function to return Continue(); instead of done. This prevents bots in Mann vs. Machine from getting stuck in an endless loop if their behavior is set to ignore the flag using "Attributes IgnoreFlag".