[TF2] Change max FOV from 90 to 130
This PR is made to increase the max field of view within Team Fortress 2 from 90 to 130.
This would help players on standard widescreen monitors be able to see as much as players on ultrawide displays, help reduce nausea, to have parity with Valve's other titles such as Left 4 Dead 2 as well as the Team Fortress modding communities typically allowed max FOV settings.