Revise "connected player" check for AOE healing
Previously, we checked for AoE healing by checking if said healing didn't provide overheal. We now instead check if a healing player is not holding a Medigun, since AoE healing is currently only done while holding the Amputator (a melee weapon). This is less likely to cause issues with modded weapons.