Added background01.vtf to replace load screen
Seems to be hard coded to use "background01" (or widescreen variant) for
loading screen - since the menu splash screen names were changed, this
reference was broken and the default TF2 load screen was being used.
Including just a background01.vmt that references a different file does
not work; likely directly references the VTF in code

This background01.vtf should be replaced with our loading screen later.