Fix ordering of hitsound calls in CBaseHLBludgeonWeapon::Swing() and play melee hitsound in CBaseHLBludgeonWeapon::Hit()
This commit properly fixes melee weapon hitsounds in Half-Life 2 and derivative mods by adding a `WeaponSound( MELEE_HIT )`/`WeaponSound( MELEE_HIT_WORLD )` call at the end of the during the hit target entity check in `CBaseHLBludgeonWeapon::Hit()`, and changes the ordering of the swing `WeaponSound` call in `CBaseHLBludgeonWeapon::Swing()` to be placed just before the hit or miss conditional blocks.

This exact hitsound fix is only PARTIALLY present in HL20 exclusively for the player stunstick ent (regular stunstick and crowbar don't work), because the fix that's included is quite literally in the wrong place!