Fix reflected ornaments shattering in the wrong color
When a Wrap Assassin's ornament is reflected by an enemy, it changes its color to that of the enemy's team. However, when it shatters, the particles it creates are of the original thrower's team color rather than the reflecting enemy's team color. This means RED ornaments can shatter into BLU shards and vice versa. This commit fixes that.