Play ornament shatter sounds more loudly to reflector when reflected
When a Wrap Assassin ornament shatters, it plays a shattering sound. This sound plays at a higher volume to the person who threw the ornament so they know when it breaks. When an enemy reflects an ornament, however, the sound still plays at a higher volume to the original thrower rather than the reflecting enemy, even though the information of when the ornament breaks is more useful to the reflector than the thrower. This commit makes the sound effect play louder to the reflector rather than the thrower when applicable.