Fix humiliation death notices not reflecting feigned deaths
In other words, this commit fixes the Holy Mackerel, Unarmed Combat, and Hot Hand displaying "x1"/"x2"/etc. messages instead of "FISH/ARM/SLAP KILL!" messages in the killfeed when a Spy feigns death with the Dead Ringer.

The short reason for why this occurs is that m_bGoingFeignDeath isn't set when these messages are triggered, which means that TF2 doesn't know that such a message coincides with a Spy feigning their death.