Replace uses of `GetOwnerEntity` with `GetThrower` for baseballs and ornaments
GetOwnerEntity returns the 'owner entity' of the baseball/ornament, which is the player that originally threw it. However, this is used in multiple places where we really want to know who is responsible for the ball's damage (including who may have last reflected it) which is returned by GetThrower.

This change should fix reflected baseball/ornament hits and kills being attributed to their original thrower rather than their reflector like other projectiles.